Add files via upload
This simulation will operate on 8-bit data and perform arithmetic and logic operations. To operate the ALU, we’ll use a 21-bit binary value as controlling signal, it demonstrates the heart of computation.
In upcoming file instruction decoder will added to this simulation. Feel free to expand upon this simulation by adding more features and instructions!